rpcclient: update rescanblockchain support

+++ btcjson/walletsvrresults.go
@@ -319,8 +319,8 @@ type ListUnspentResult struct {
 
 // RescanBlockchainResult models the data returned from the rescanblockchain command.
 type RescanBlockchainResult struct {
-	StartHeight int64 `json:"start_height"`
-	StoptHeight int64 `json:"stop_height"`
+	StartHeight int32 `json:"start_height"`
+	StoptHeight int32 `json:"stop_height"`
 }
 
 // SignRawTransactionError models the data that contains script verification
